1. Deletes results messages
2. Download speed now visible
3. postproc.py checks for gclone success before deleting

+++ postproc.py
@@ -60,7 +60,7 @@ if category == "tv":
     RCLONE_UPLOAD_DIRECTORY = f'{directory.split("/")[-2]}/{directory.split("/")[-1]}'
 DRIVE_UPLOAD_DIRECTORY = f"{RCLONE_REMOTE_NAME}:{RCLONE_DIRECTORY_NAME}/{RCLONE_UPLOAD_DIRECTORY}"
 
-rclone_command = f"gclone copy -v --stats=1s --stats-one-line --drive-chunk-size=256M --fast-list --transfers=1 --exclude _UNPACK_*/** --exclude _FAILED_*/** --exclude *.rar \"{directory}\" \"{DRIVE_UPLOAD_DIRECTORY}\" "
+rclone_command = f"gclone copy -v --stats=1s --stats-one-line --drive-chunk-size=256M --fast-list --transfers=1 --exclude _UNPACK_*/** --exclude _FAILED_*/** --exclude *.rar --exclude *.par2 \"{directory}\" \"{DRIVE_UPLOAD_DIRECTORY}\" "
 
 logging.basicConfig(
     level=logging.INFO,
@@ -177,7 +177,7 @@ def webhook_notification_embed(message:str,**kwargs):
     sys.exit(1)
 
 
-def run_command(command):
+def run_command(command) -> bool:
     with subprocess.Popen(
         command, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, shell=True
     ) as proc:
@@ -191,8 +191,9 @@ def run_command(command):
                 LOGGER(__name__).info(f"Uploading to drive: {output}")
 
             if output == "" and proc.poll() is not None:
-                LOGGER(__name__).info("File has been successfully uploaded to gdrive.")
-                break
+                return True
+            else: 
+                return False
 
 
 if re.search(r"(http|https)", jobname):
@@ -205,6 +206,7 @@ reasons = {
     "3": "Failed unpack / verification",
 }
 
+LOGGER(__name__).info(f"postprocstatus: {postprocstatus}")
 
 if str(postprocstatus) in reasons:
     reason = reasons[postprocstatus]
@@ -213,47 +215,51 @@ if str(postprocstatus) in reasons:
     sys.exit(1)
 
 
-run_command(rclone_command)
+gclone_status = run_command(rclone_command)
 
-# deleting file from local drive.
-contains_prefix = False
-prefix = "_UNPACK_"
-for folder_name in os.listdir(directory):
-    if folder_name.startswith(prefix) and os.path.isdir(os.path.join(directory, folder_name)):
-        LOGGER(__name__).info(f"Folder starting with {prefix} found: {folder_name}")
-        contains_prefix = True
-    else:
-        path = os.path.join(directory, folder_name)
-        if os.path.isfile(path):
-            os.unlink(path)
-            LOGGER(__name__).info(f"File deleted: {path}")
-        elif os.path.isdir(path):
-            shutil.rmtree(path)
-            LOGGER(__name__).info(f"Directory deleted: {path}")
+if gclone_status:
+    LOGGER(__name__).info("File has been successfully uploaded to gdrive.")
+    # deleting file from local drive.
+    contains_prefix = False
+    prefix = "_UNPACK_"
+    for folder_name in os.listdir(directory):
+        if folder_name.startswith(prefix) and os.path.isdir(os.path.join(directory, folder_name)):
+            LOGGER(__name__).info(f"Folder starting with {prefix} found: {folder_name}")
+            contains_prefix = True
+        else:
+            path = os.path.join(directory, folder_name)
+            if os.path.isfile(path):
+                os.unlink(path)
+                LOGGER(__name__).info(f"File deleted: {path}")
+            elif os.path.isdir(path):
+                shutil.rmtree(path)
+                LOGGER(__name__).info(f"Directory deleted: {path}")
 
-if not contains_prefix:
-  shutil.rmtree(directory)
-  LOGGER(__name__).info(f"Directory deleted: {directory}")
-try:
-    file_size = os.environ["SAB_BYTES_DOWNLOADED"]
-    file_size = get_readable_bytes(int(file_size))
-except:
-    file_size = "N/A"
-
-
-drive_link = ""
-if SHOW_DRIVE_LINK:
-    drive_link = subprocess.check_output(
-        ["rclone", "link", f"{DRIVE_UPLOAD_DIRECTORY}"]).decode("utf-8")
-
-    if "drive.google.com" not in drive_link:
-        drive_link = "Something went wrong!"
-
-    print(drive_link)
-    drive_link = f'[Drive Link]({encode_link(drive_link,"SecretBot")})'
+    if not contains_prefix:
+      shutil.rmtree(directory)
+      LOGGER(__name__).info(f"Directory deleted: {directory}")
+    try:
+        file_size = os.environ["SAB_BYTES_DOWNLOADED"]
+        file_size = get_readable_bytes(int(file_size))
+    except:
+        file_size = "N/A"
 
 
-notification_message = (
-    f"`📁 {jobname}`\n\n{file_size} | Success |{drive_link}")
+    drive_link = ""
+    if SHOW_DRIVE_LINK:
+        drive_link = subprocess.check_output(
+            ["rclone", "link", f"{DRIVE_UPLOAD_DIRECTORY}"]).decode("utf-8")
 
-webhook_notification(message=notification_message)
+        if "drive.google.com" not in drive_link:
+            drive_link = "Something went wrong!"
+
+        print(drive_link)
+        drive_link = f'[Drive Link]({encode_link(drive_link,"SecretBot")})'
+
+
+    notification_message = (
+        f"`📁 {jobname}`\n\n{file_size} | Success |{drive_link}")
+
+    webhook_notification(message=notification_message)
+else:
+    LOGGER(__name__).info("File failed to upload file to drive.")
